refactor(playback): split activation-hold static family into GhostPlaybackEngine.ActivationHold.cs
